requirements.txt
discord.py>=2.7,<3
loguru>=0.7,<1
matplotlib>=3.8,<4
playwright>=1.58,<2
pyotp>=2.9,<3
pyserial>=3.5,<4
qq-botpy>=1.2,<2
requests>=2.32,<3
rich>=14,<15
textual>=0.80,<9
